+++ mobile/android/android-components/components/feature/toolbar/src/test/java/mozilla/components/feature/toolbar/ToolbarAutocompleteFeatureTest.kt
@@ -104,7 +104,11 @@ class ToolbarAutocompleteFeatureTest {
             fail()
         }
 
-        override fun addEditAction(action: Toolbar.Action) {
+        override fun addEditActionStart(action: Toolbar.Action) {
+            fail()
+        }
+
+        override fun addEditActionEnd(action: Toolbar.Action) {
             fail()
         }